Made some new classes: Point and GameObject
Point is a class for points that can compare == and != and also implements add and subtract.

Made a GameObject class that all game objects inherit from which have a Point position, a direction, a type and a few other things.

Converted all game objects (players and prizes) over to inherit from this class and updated all code that interfaced with these game objects.